requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
Django==5.1.2
django-autoslug==1.9.9
django-crispy-forms==2.3
mysql==0.0.3
mysqlclient==2.2.4
sqlparse==0.5.1
tzdata==2024.2
